Make minor fix in tag repository

+++ api/getUserTags.go
@@ -3,5 +3,5 @@ package api
 import "github.com/VitaliiHurin/go-newsfeed/entity"
 
 func (a *API) GetUserTags(user *entity.User) ([]*entity.Tag, error) {
-	return a.tags.GetByUser(int64(user.ID))
+	return a.tags.GetByUser(user.ID)
 }
